refactor: move navigation components
