+++ Components/Navigation/IDockable.cs
@@ -1,3 +1,5 @@
+using Spacelancer.Components.Navigation.Software;
+
 namespace Spacelancer.Components.Navigation;
 
 public interface IDockable : INavigable
