+++ Components/Navigation/Software/SystemAutoNavigation.cs
@@ -2,7 +2,7 @@ using System;
 using Godot;
 using Serilog;
 
-namespace Spacelancer.Components.Navigation;
+namespace Spacelancer.Components.Navigation.Software;
 
 public class SystemAutoNavigation : AutomatedNavigation
 {
