+++ Scenes/SpaceShips/PlayerDebug.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using Godot;
 using Spacelancer.Components.Navigation;
+using Spacelancer.Components.Navigation.Software;
 
 namespace Spacelancer.Scenes.SpaceShips;
