+++ Scenes/SpaceShips/PlayerEquipment.cs
@@ -3,6 +3,7 @@ using Spacelancer.Components.Equipment.Detection;
 using Spacelancer.Components.Equipment.Storage;
 using Spacelancer.Components.Navigation;
 using Spacelancer.Components.Navigation.Computers;
+using Spacelancer.Components.Navigation.Software;
 using Spacelancer.Controllers;
 using Spacelancer.Economy;
