+++ Scenes/Transitions/LaneNode.cs
@@ -2,6 +2,7 @@ using Godot;
 using Serilog;
 using Spacelancer.Components.Equipment.Detection;
 using Spacelancer.Components.Navigation;
+using Spacelancer.Components.Navigation.Software;
 
 namespace Spacelancer.Scenes.Transitions;
